Revert "Use BoLD contracts to test delay buffer"
This reverts commit b0d586cb4dcc13d414de86eb35958f5bb04be54a.
